RavenDB-16201 WIP
